# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 数据分析 Agent 本地服务 API
  version: 1.0.0
  description: |
    数据分析 Agent 的本地 FastAPI 服务契约。前端据此生成 TypeScript 类型与 MSW Mock，
    后端据此实现。详见 docs/02-架构.md。

    统一约定：
    - 所有端点带 /api/v1 前缀。
    - 统一响应信封 { code, data, message }，code=0 表示成功。
    - 分页统一 { items, total, page, limit, hasMore } 置于 data 内。
    - 认证为「仅本机回环 + Origin 白名单」；标注 x-local-confirm 的端点需
      请求头 X-Local-Confirm: true。
    - 禁止静默降级：任何能力降级必须在响应字段上可见。
  contact:
    name: 高见远（首席架构师）
servers:
  - url: http://127.0.0.1:8000
    description: 本地服务（生产由本服务托管前端 dist，单端口）
...
paths:
...
  /api/v1/llm/routes:
    get:
      tags: [llm]
      summary: 查看 taskKey 到模型的路由表
      operationId: getRoutes
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/LlmRoute'
    put:
      tags: [llm]
      summary: 覆盖式更新路由表
      operationId: putRoutes
      parameters:
        - $ref: '#/components/parameters/LocalConfirm'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/LlmRoute'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/v1/memories:
    get:
      tags: [memories]
      summary: 长期记忆列表
      operationId: listMemories
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: scope
          in: query
          schema: { type: string, enum: [global, conversation, dataset] }
        - name: conversationId
          in: query
          schema: { type: string, format: uuid }
        - name: kind
          in: query
          schema: { type: string, enum: [fact, preference, insight] }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageData'
                          - properties:
                              items:
                                type: array
                                items: { $ref: '#/components/schemas/Memory' }
    post:
      tags: [memories]
      summary: 手动写入记忆
      operationId: createMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Memory'
  /api/v1/memories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [memories]
      summary: 修改内容或权重
      operationId: updateMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string, maxLength: 4000 }
                weight: { type: number, minimum: 0, maximum: 10 }
                expiresAt: { type: string, format: date-time, nullable: true }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [memories]
      summary: 删除记忆
      operationId: deleteMemory
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }

  # ────────────────────────── knowledge ──────────────────────────
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: 资源 UUID
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    LocalConfirm:
      name: X-Local-Confirm
      in: header
      required: true
      schema: { type: string, enum: ['true'] }
      description: 写密钥或不可逆操作需携带。缺失时返回 403 且 code=4010。
...
  schemas:
    ApiResponse:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          description: 0 表示成功；非 0 见 docs/02-架构.md §6.8 业务错误码
          example: 0
        data:
          nullable: true
        message:
          type: string
          default: ''
    PageData:
      type: object
      required: [items, total, page, limit, hasMore]
      properties:
        items:
          type: array
          items: {}
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        hasMore: { type: boolean }
    PageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
...
    TaskKey:
      type: string
      enum: [plan_analysis, narrate_insights, parse_question, text_to_sql, memory_summarize, chunk_enrich, report_narrate, chat_reply]
...
    LlmRoute:
      type: object
      required: [taskKey, primaryProviderId, primaryModel]
      properties:
        taskKey: { $ref: '#/components/schemas/TaskKey' }
        primaryProviderId: { type: string, format: uuid }
        primaryModel: { type: string }
        fallback:
          type: array
          items:
            type: object
            properties:
              providerId: { type: string, format: uuid }
              model: { type: string }
...
    Memory:
      type: object
      properties:
        id: { type: string, format: uuid }
        scope: { type: string, enum: [global, conversation, dataset] }
        conversationId: { type: string, format: uuid, nullable: true }
        datasetId: { type: string, format: uuid, nullable: true }
        kind: { type: string, enum: [fact, preference, insight] }
        content: { type: string }
        weight: { type: number }
        sourceMessageId: { type: string, format: uuid, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MemoryCreate:
      type: object
      required: [scope, kind, content]
      properties:
        scope: { type: string, enum: [global, conversation, dataset] }
        conversationId: { type: string, format: uuid, nullable: true }
        datasetId: { type: string, format: uuid, nullable: true }
        kind: { type: string, enum: [fact, preference, insight] }
        content: { type: string, minLength: 1, maxLength: 4000 }
        weight: { type: number, default: 1, minimum: 0, maximum: 10 }
...